extra_features/requirements.txt
bark_ssg==1.3.4
boto3==1.29.4
google_api_python_client==2.107.0
google_auth_oauthlib==1.0.0
ipython==7.31.1
numpy==1.24.4
numpy==1.21.5
openai==1.6.1
protobuf==3.20.3
protobuf==3.12.4
pymilvus==2.3.3
scipy==1.8.0
sentence_transformers==2.2.2
